+++ test/integration/any_element_city_of_london.py
@@ -6,7 +6,7 @@ assert __name__ == "__main__"
 validate_elements_in_result_set(
     code="""
 [timeout:60];
-area[name="Carabanchel"][boundary=administrative];
+area[wikipedia="en:City of London"];
 nwr(area);
 out geom;
 """