+++ test/integration/any_route_city_of_london.py
@@ -6,7 +6,7 @@ assert __name__ == "__main__"
 validate_routes_in_result_set(
     code="""
 [timeout:180];
-area[name="Carabanchel"][boundary=administrative];
+area[wikipedia="en:City of London"];
 rel(area)[type=route]->.routes;
 """
 )